NickAkhmetov/HMP-515 Always display `Launch Workspace` button (#3342)
