+++ CHANGELOG-HMP-515.md
@@ -1,0 +1,1 @@
+- Allow reopening closed workspace tabs without needing to relaunch the entire session.
