--- context/app/static/js/components/workspaces/WorkspaceLaunchStopButton/index.ts
@@ -1,3 +1,0 @@
-import WorkspaceLaunchStopButton from './WorkspaceLaunchStopButton';
-
-export default WorkspaceLaunchStopButton;
