+++ context/app/static/js/components/workspaces/WorkspaceLaunchStopButtons/index.ts
@@ -1,0 +1,3 @@
+import WorkspaceLaunchStopButtons from './WorkspaceLaunchStopButtons';
+
+export default WorkspaceLaunchStopButtons;
